refactor: classes; feat: project Popup
