+++ src/components/card.js
@@ -1,96 +1,91 @@
-import {userId} from '../index.js';
-import {card, gallery, imageFull, descriptionFull, imagePopup, openPopup} from './utils.js';
-import {config, deleteCard, likeCard, dislikeCard} from './Api.js';
+import { userId } from '../index.js'
+export default class Card {
+  // liker, deleter, imageOpener - функции-колбеки, используют запросы (т.е. методы)  класса api (объявим в index.js)
+  constructor(object, selectorTemplate, liker, deleter, imageOpener){
+    // из object нам понадобятся: _id, name, link, likes, owner._id
+    this._object = object;
+    this._selectorTemplate = selectorTemplate;
+    this._liker = liker;
+    this._deleter = deleter;
+    this._imageOpener = imageOpener;
+  }
 
-// // функция "вставить карточку" в начало галереи
-export function pasteCard(card) {
-  gallery.prepend(card);
-}
+  _getElement() {
+    // клонирует и возвращает разметку формы
+    const cardElement = document
+      .querySelector(this._selectorTemplate)
+      .content
+      .querySelector('.card')
+      .cloneNode(true);
 
-//функция разворачивания картинки
-export function openImage(name, link) {
-  imageFull.src = link; // при открытии заполняем значение поля указанным на странице
-  imageFull.alt = name;
-  descriptionFull.textContent = name;
+    return cardElement;
+  }
 
-  openPopup(imagePopup); // и показываем попап с заполненным содержимым
-}
+  getId() { // мб пригодится если делать подтверждение удаления
+    // id карточки
+    return this._object._id;
+  }
 
-// // функционал карточки (лайки, удаление, разворот картинки)
-export function addFunctional(inpName, inpLink, cardObj, imageEl, likeNumber, likeBtn, deleteBtn) {
-  // навешиваем возможность "лайкать"
-  likeBtn.addEventListener('click', function(evt) {
+  _isLiked() {
+    // если среди лайков есть мой (пользователя) - вернет true
+    return this._object.likes.some(item => item._id === userId)
+  }
 
-// если лайк уже стоял - убираем его и обновляем инфо на сервере
-    if (evt.target.classList.contains('js-active')) {
-      dislikeCard(config, cardObj)
-        .then((data) => {
-          evt.target.classList.toggle('js-active'); // меняем цвет сердечка на противоположный
-          likeNumber.textContent = data.likes.length; // обновляем количество лайков
-        })
-        .catch((err) => {
-          console.log(err);
-        });
-    // иначе - ставим (и тоже обновляем)
+  _countLikes() {
+    // меняет счетчик лайков в разметке и отвечает за покраску сердечка
+    this._likeNumber.textContent = this._object.likes.length;
+
+    if (this._isLiked()) {
+      this._likeBtn.classList.add('js-active');
     } else {
-      likeCard(config, cardObj)
-        .then((data) => {
-          evt.target.classList.toggle('js-active'); // меняем цвет сердечка на противоположный
-          likeNumber.textContent = data.likes.length; // обновляем количество лайков
-        })
-        .catch((err) => {
-          console.log(err);
-        });
+      this._likeBtn.classList.remove('js-active');
     }
-  });
-
-  // навешиваем возможность удалять
-  deleteBtn.addEventListener('click', function(evt) {
-    const cardItem = evt.target.closest('.card');
-
-    // удаляем карточку с сервера
-    deleteCard(config, cardObj)
-      .then(() => {
-        cardItem.remove();
-      })
-      .catch((err) => {
-        console.log(err);
-      })
-  });
-
-  // навешиваем возможность разворачивать картинку на каждую картинку
-  imageEl.addEventListener('click', function() {
-    openImage(inpName, inpLink); // используем функцию
-  });
-}
-
-// создание карточки
-export function makeCard(object) { // принимает на вход объект
-  const cardCopy = card.cloneNode(true); // клонируем содержимое шаблона
-  // сразу находим все нужные элементы для дальнейшего исп-я (в т.ч. в функции addFunctional)
-  const imageEl = cardCopy.querySelector('.card__image');
-  const likeNumber = cardCopy.querySelector('.card__like-number');
-  const likeBtn = cardCopy.querySelector('.btn_el_like');
-  const deleteBtn = cardCopy.querySelector('.btn_el_delete');
-  // далее заполняем шаблон
-  imageEl.alt = object.name;
-  imageEl.src = object.link;
-  cardCopy.querySelector('.card__title').textContent = object.name;
-  cardCopy.querySelector('.card__like-number').textContent = object.likes.length; // количество лайков
-
-  addFunctional(object.name, object.link, object, imageEl, likeNumber, likeBtn, deleteBtn); // добавляем карточке функционал
-
-  // если карточка не моя, убираем иконку удаления
-  if (object.owner._id !== userId) {
-    cardCopy.querySelector('.btn_el_delete').remove();
   }
 
-  // если среди лайков есть мой, закрашиваем сердечко
-  if (object.likes.some(item => {
-    return item._id === userId
-  })) {
-    cardCopy.querySelector('.btn_el_like').classList.add('js-active');
+  _removeDelBtn() {
+    // если карточка не моя, убираем иконку удаления
+    if (this._object.owner._id !== userId) {
+      this._cardCopy.querySelector('.btn_el_delete').remove();
+    }
   }
 
-  return cardCopy;
+  _addFunctional() {
+    // используем колбеки (будем передавать запросы Api)
+    this._likeBtn.addEventListener('click', (evt) => this._liker(this._object, evt, this._likeNumber));
+    this._deleteBtn.addEventListener('click', () => this._deleter(this._object, this._deleteBtn));
+    this._imageEl.addEventListener('click', () => this._imageOpener(this._object));
+  }
+
+// создает и возвращает полностью готовую карточку
+  makeCard() {
+    this._cardCopy = this._getElement(); // клонируем содержимое шаблона
+    // сразу находим все нужные элементы для дальнейшего исп-я (в т.ч. в функции addFunctional)
+    this._imageEl = this._cardCopy.querySelector('.card__image');
+    this._likeNumber = this._cardCopy.querySelector('.card__like-number');
+    this._likeBtn = this._cardCopy.querySelector('.btn_el_like');
+    this._deleteBtn = this._cardCopy.querySelector('.btn_el_delete');
+
+    // далее заполняем шаблон
+    this._imageEl.alt = this._object.name;
+    this._imageEl.src = this._object.link;
+    this._cardCopy.querySelector('.card__title').textContent = this._object.name;
+
+    // добавляем карточке функционал (вешаем слушатели)
+    this._addFunctional();
+
+    // если карточка не моя, убираем иконку удаления
+    this._removeDelBtn();
+
+    // если среди лайков есть мой, закрашиваем сердечко
+    this._countLikes();
+
+    return this._cardCopy;
+  }
+
+  delete() {
+    // удаляет карточку
+    this._cardCopy.remove();
+  }
 }
+
+
